Revert "Beginning to add Tagging"
This reverts commit c3547415c7dce6e00a40d95db162797dec10a5b0.
