+++ content/modules/dev/ui.docView.video.js
@@ -767,18 +767,6 @@ var ytMetadataCallbacks = jQuery.Deferred();
 			case "metronome":
 				console.log("Tick");
 				if (!self._ignoremetronome){
-					//is there an annotation over the upcoming metronom period
-					var p0 = Math.floor(evt.value * self.SEC_MULT_FACTOR);
-					var p1 = p0 + self.SEC_MULT_FACTOR * self.T_METRONOME; 
-					var locs = self._model.get("location", {page__in: [p0, p1]});
-					if (!locs.is_empty()){
-           					var firstlocid = String(locs.first().ID);
-            					if (firstlocid !== self._id_location){
-							console.log("New Thread");
-							$.concierge.trigger({type:"select_thread", value: firstlocid});
-            					}
-					}
-
 					NB_vid.methods.updatePlayerInfo();
 					NB_vid.methods.updateProgressbar();
 				}
