+++ content/modules/dev/ui.notepaneView.video.js
@@ -114,16 +114,16 @@
 			break;
             case "metronome": 
 			//is there an annotation over the upcoming metronom period
-//			var p0 = Math.floor(evt.value * self.SEC_MULT_FACTOR);
-//			var p1 = p0 + self.SEC_MULT_FACTOR * self.T_METRONOME; 
-//			var locs = self._model.get("location", {page__in: [p0, p1]});
-//			if (!locs.is_empty()){
-//            var firstlocid = String(locs.first().ID);
-//            if (firstlocid !== self._id_location){
-//				console.log("New Thread");
-//				$.concierge.trigger({type:"select_thread", value: firstlocid});
-//            }
-//			}
+			var p0 = Math.floor(evt.value * self.SEC_MULT_FACTOR);
+			var p1 = p0 + self.SEC_MULT_FACTOR * self.T_METRONOME; 
+			var locs = self._model.get("location", {page__in: [p0, p1]});
+			if (!locs.is_empty()){
+            var firstlocid = String(locs.first().ID);
+            if (firstlocid !== self._id_location){
+				console.log("New Thread");
+				$.concierge.trigger({type:"select_thread", value: firstlocid});
+            }
+			}
 			break;
             }	
 		}
